requirements.txt
requests==2.32.3
python-dotenv==1.0.1
click==8.1.7
